construction_pmis/requirements.txt
frappe-sdk>=1.0.0
erpnext>=14.0.0 # Assuming ERPNext v14 is the target as per requirements
